+++ debug_toolbar/__init__.py
@@ -3,7 +3,7 @@ __all__ = ["VERSION"]
 
 # Do not use pkg_resources to find the version but set it here directly!
 # see issue #1446
-VERSION = "3.2"
+VERSION = "3.2.1"
 
 # Code that discovers files or modules in INSTALLED_APPS imports this module.