+++ debug_toolbar/panels/history/views.py
@@ -1,15 +1,16 @@
 from django.http import HttpResponseBadRequest, JsonResponse
 from django.template.loader import render_to_string
 
-from debug_toolbar.decorators import require_show_toolbar
+from debug_toolbar.decorators import require_show_toolbar, signed_data_view
 from debug_toolbar.panels.history.forms import HistoryStoreForm
 from debug_toolbar.toolbar import DebugToolbar
 
 
 @require_show_toolbar
-def history_sidebar(request):
+@signed_data_view
+def history_sidebar(request, verified_data):
     """Returns the selected debug toolbar history snapshot."""
-    form = HistoryStoreForm(request.GET)
+    form = HistoryStoreForm(verified_data)
 
     if form.is_valid():
         store_id = form.cleaned_data["store_id"]
@@ -32,9 +33,10 @@ def history_sidebar(request):
 
 
 @require_show_toolbar
-def history_refresh(request):
+@signed_data_view
+def history_refresh(request, verified_data):
     """Returns the refreshed list of table rows for the History Panel."""
-    form = HistoryStoreForm(request.GET)
+    form = HistoryStoreForm(verified_data)
 
     if form.is_valid():
         requests = []